refactor(quiz): :recycle: moved all quiz types into services/types module
